Set bridges

ConstraintProgrammingExtensions.Bridges.add_all_set_bridges — Function

add_all_set_bridges(bridged_model, ::Type{T}) where {T}

Add all set bridges defined in the Bridges submodule to bridged_model. The coefficient type used is T.

ConstraintProgrammingExtensions.Bridges.AbsoluteValue2MILPBridge — Type

Bridges CP.AbsoluteValue to linear constraints and integer constraints.

The implemented model is the most generic one, so that the absolute value always has a well-defined value. This requires the use of a binary variable. In many cases, this is not necessary, and simpler models could be used, but checking this automatically would require access to the whole model.

Based on Mosek's modelling cookbook.

YALMIP implements a similar formulation, with big-M and small-m values computed based on the bounds for each variable.

ConstraintProgrammingExtensions.Bridges.AllDifferent2DifferentFromBridge — Type

Bridges MOI.AllDifferent to a series of CP.DifferentFrom.

ConstraintProgrammingExtensions.Bridges.AllDifferentExceptConstants2ConjunctionDisjunctionBridge — Type

Bridges CP.AllDifferentExceptConstants to a series of CP.Disjunction of CP.DifferentFrom and CP.Domain.

ConstraintProgrammingExtensions.Bridges.AllDifferentExceptConstants2ReificationBridge — Type

Bridges CP.AllDifferentExceptConstants to reifications.

ConstraintProgrammingExtensions.Bridges.AllEqual2EqualToBridge — Type

Bridges CP.AllEqual to a series of CP.EqualTo.

ConstraintProgrammingExtensions.Bridges.ArgumentMaximumAmong2MILPBridge — Type

Bridges CP.ArgumentMaximumAmong to MILP formulations, by the means of big-M constraints.

ConstraintProgrammingExtensions.Bridges.ArgumentMinimumAmong2MILPBridge — Type

Bridges CP.ArgumentMinimumAmong to MILP formulations, by the means of big-M constraints.

ConstraintProgrammingExtensions.Bridges.BinPacking2MILPBridge — Type

Bridges CP.BinPacking to a MILP by creating binary variables for the bin assignment and MILP constraints.

ConstraintProgrammingExtensions.Bridges.FixedCapacityBinPacking2BinPackingBridge — Type

Bridges CP.BinPacking{CP.FIXED_CAPACITY_BINPACKING} to CP.BinPacking{CP.NO_CAPACITY_BINPACKING} by adding constraints on the capacity variables.

ConstraintProgrammingExtensions.Bridges.FixedCapacityBinPacking2VariableCapacityBinPackingBridge — Type

Bridges CP.BinPacking{CP.FIXED_CAPACITY_BINPACKING, T} to CP.BinPacking{CP.VARIABLE_CAPACITY_BINPACKING, T} by creating capacity variables.

ConstraintProgrammingExtensions.Bridges.VariableCapacityBinPacking2BinPackingBridge — Type

Bridges CP.BinPacking{CP.VARIABLE_CAPACITY_BINPACKING, T} to CP.BinPacking{CP.NO_CAPACITY_BINPACKING, T} by adding constraints on the capacity variables.

ConstraintProgrammingExtensions.Bridges.GlobalCardinalityFixedClosed2GlobalCardinalityFixedOpenBridge — Type

Bridges CP.GlobalCardinality{CP.FIXED_COUNTED_VALUES, CP.CLOSED_COUNTED_VALUES, T} to CP.GlobalCardinality{CP.FIXED_COUNTED_VALUES, CP.OPEN_COUNTED_VALUES, T.

ConstraintProgrammingExtensions.Bridges.GlobalCardinalityVariableClosed2GlobalCardinalityVariableOpenBridge — Type

Bridges GlobalCardinality{CP.VARIABLE_COUNTED_VALUES, CP.CLOSED_COUNTED_VALUES, T} to GlobalCardinality{CP.VARIABLE_COUNTED_VALUES, CP.OPEN_COUNTED_VALUES, T}.

ConstraintProgrammingExtensions.Bridges.Conjunction2ReificationBridge — Type

Bridges CP.Conjunction to reification.

ConstraintProgrammingExtensions.Bridges.Count2ReificationBridge — Type

Bridges CP.Count to reification.

ConstraintProgrammingExtensions.Bridges.CountCompare2CountBridge — Type

Bridges CP.CountCompare to CP.Count.

ConstraintProgrammingExtensions.Bridges.Decreasing2LPBridge — Type

Bridges CP.Decreasing to linear constraints.

ConstraintProgrammingExtensions.Bridges.DifferentFrom2PseudoMILPBridge — Type

Bridges CP.DifferentFrom to linear constraints (including, possibly, strict inequalities). This constraint adds one variable to store the absolute value of the difference, and constrains it to be nonzero.

For AbstractFloat arguments (like Float64): equivalent to abs(x) > 0.0, i.e. a Strictly(GreaterThan(0.0)).

ConstraintProgrammingExtensions.Bridges.IndicatorDifferentFrom2PseudoMILPBridge — Type

Bridges MOI.Indicator{A, CP.DifferentFrom} to linear constraints (including, possibly, strict inequalities). This constraint adds one variable to store the absolute value of the difference, and uses it for the indicator.

For AbstractFloat arguments (like Float64): equivalent to abs(x) > 0.0, i.e. a Strictly(GreaterThan(0.0)).

ConstraintProgrammingExtensions.Bridges.ReificationDifferentFrom2IndicatorBridge — Type

Bridges CP.Reification{CP.DifferentFrom} to indicator constraints, both with equality and inequalities (CP.DifferentFrom).

ConstraintProgrammingExtensions.Bridges.ReificationDifferentFrom2MILPBridge — Type

Bridges CP.Reification{CP.DifferentFrom} to MILP constraints.

ConstraintProgrammingExtensions.Bridges.Disjunction2ReificationBridge — Type

Bridges CP.Disjunction to reification.

ConstraintProgrammingExtensions.Bridges.Domain2MILPBridge — Type

Bridges CP.Domain to MILP by adding one binary variable per possible combination.

ConstraintProgrammingExtensions.Bridges.DoublyLexicographicallyLessThan2LexicographicallyLessThanBridge — Type

Bridges CP.DoublyLexicographicallyLessThan to CP.LexicographicallyLessThan.

ConstraintProgrammingExtensions.Bridges.DoublyLexicographicallyGreaterThan2LexicographicallyGreaterThanBridge — Type

Bridges CP.DoublyLexicographicallyGreaterThan to CP.LexicographicallyGreaterThan.

ConstraintProgrammingExtensions.Bridges.Element2MILPBridge — Type

Bridges CP.Element to MILP constraints by using a unary encoding of the index in the array.

ConstraintProgrammingExtensions.Bridges.ElementVariableArray2MILPBridge — Type

Bridges CP.ElementVariableArray to MILP constraints by using a unary encoding of the index in the array.

ConstraintProgrammingExtensions.Bridges.ReificationEqualTo2IndicatorBridge — Type

Bridges CP.Reification{MOI.EqualTo} to indicator constraints, both with equality and inequalities (CP.DifferentFrom).

ConstraintProgrammingExtensions.Bridges.ReificationEqualTo2MILPBridge — Type

Bridges CP.Reification{MOI.EqualTo} to MILP constraints.

ConstraintProgrammingExtensions.Bridges.ReificationGreaterThan2IndicatorBridge — Type

Bridges CP.Reification{MOI.GreaterThan} to indicator constraints with (strict) inequalities.

ConstraintProgrammingExtensions.Bridges.ReificationGreaterThan2MILPBridge — Type

Bridges CP.Reification{MOI.GreaterThan} to MILP constraints.

ConstraintProgrammingExtensions.Bridges.GlobalCardinalityFixedOpen2CountBridge — Type

Bridges CP.GlobalCardinality to CP.Count.

ConstraintProgrammingExtensions.Bridges.GlobalCardinalityFixedOpen2GlobalCardinalityVariableOpenBridge — Type

Bridges CP.GlobalCardinality to CP.GlobalCardinalityVariable.

ConstraintProgrammingExtensions.Bridges.GlobalCardinalityVariableOpen2CountBridge — Type

Bridges CP.GlobalCardinality{CP.VARIABLE_COUNTED_VALUES, CP.OPEN_COUNTED_VALUES, T} to CP.Count.

ConstraintProgrammingExtensions.Bridges.IfThenElse2ImplicationBridge — Type

Bridges CP.IfThenElse to CP.Implication.

ConstraintProgrammingExtensions.Bridges.IfThenElse2ReificationBridge — Type

Bridges CP.IfThenElse to reification.

ConstraintProgrammingExtensions.Bridges.Implication2ReificationBridge — Type

Bridges CP.Implication to reification.

ConstraintProgrammingExtensions.Bridges.Increasing2LPBridge — Type

Bridges CP.Increasing to linear constraints.

ConstraintProgrammingExtensions.Bridges.Inverse2ReificationBridge — Type

Bridges CP.Inverse to reification.

ConstraintProgrammingExtensions.Bridges.Knapsack2MILPBridge — Type

Bridges CP.Knapsack{KCT, KVT} to a MILP by adding the corresponding MILP constraint.

ConstraintProgrammingExtensions.Bridges.Knapsack2VariableCapacityKnapsackBridge — Type

Bridges CP.Knapsack{CP.FIXED_CAPACITY_KNAPSACK, CP.UNVALUED_KNAPSACK} to CP.Knapsack{CP.VARIABLE_CAPACITY_KNAPSACK, CP.UNVALUED_KNAPSACK} by creating capacity variables.

ConstraintProgrammingExtensions.Bridges.ValuedKnapsack2KnapsackBridge — Type

Bridges CP.Knapsack{CP.FIXED_CAPACITY_KNAPSACK, CP.VALUED_KNAPSACK, T} to CP.Knapsack{CP.FIXED_CAPACITY_KNAPSACK, CP.UNVALUED_KNAPSACK, T} by creating a value constraint.

ConstraintProgrammingExtensions.Bridges.VariableCapacityValuedKnapsack2VariableCapacityKnapsackBridge — Type

Bridges CP.Knapsack{CP.VARIABLE_CAPACITY_KNAPSACK, CP.VALUED_KNAPSACK, } to CP.Knapsack{CP.VARIABLE_CAPACITY_KNAPSACK, CP.UNVALUED_KNAPSACK, T} by creating a value constraint.

ConstraintProgrammingExtensions.Bridges.ReificationLessThan2IndicatorBridge — Type

Bridges CP.Reification{MOI.LessThan} to indicator constraints with (strict) inequalities.

ConstraintProgrammingExtensions.Bridges.ReificationLessThan2MILPBridge — Type

Bridges CP.Reification{MOI.LessThan} to MILP constraints.

ConstraintProgrammingExtensions.Bridges.LexicographicallyGreaterThan2IndicatorBridge — Type

Bridges CP.LexicographicallyGreaterThan to indicators.

ConstraintProgrammingExtensions.Bridges.LexicographicallyLessThan2IndicatorBridge — Type

Bridges CP.LexicographicallyLessThan to indicators.

ConstraintProgrammingExtensions.Bridges.StrictlyDecreasing2LPBridge — Type

Bridges CP.Strictly{CP.Decreasing} to linear constraints.

ConstraintProgrammingExtensions.Bridges.DoublyStrictlyLexicographicallyLessThan2StrictlyLexicographicallyGreaterThanBridge — Type

Bridges CP.Strictly{CP.DoublyLexicographicallyGreaterThan} to CP.Strictly{CP.LexicographicallyGreaterThan}.

ConstraintProgrammingExtensions.Bridges.DoublyStrictlyLexicographicallyLessThan2StrictlyLexicographicallyLessThanBridge — Type

Bridges CP.Strictly{CP.DoublyLexicographicallyLessThan} to CP.Strictly{CP.LexicographicallyLessThan}.

ConstraintProgrammingExtensions.Bridges.StrictlyIncreasing2LPBridge — Type

Bridges CP.Strictly{CP.Increasing} to linear constraints.

ConstraintProgrammingExtensions.Bridges.StrictlyLexicographicallyGreaterThan2IndicatorBridge — Type

Bridges CP.Strictly{CP.LexicographicallyGreaterThan} to indicators.

ConstraintProgrammingExtensions.Bridges.StrictlyLexicographicallyLessThan2IndicatorBridge — Type

Bridges CP.Strictly{CP.LexicographicallyLessThan} to indicators.

ConstraintProgrammingExtensions.Bridges.Strictly2LPBridge — Type

Bridges CP.Strictly to linear constraints.

ConstraintProgrammingExtensions.Bridges.MaximumAmong2MILPBridge — Type

Bridges CP.MaximumAmong to MILP formulations, by the means of big-M constraints.

ConstraintProgrammingExtensions.Bridges.MinimumAmong2MILPBridge — Type

Bridges CP.MinimumAmong to MILP formulations, by the means of big-M constraints.

ConstraintProgrammingExtensions.Bridges.NonOverlappingOrthotopes2DisjunctionLPBridge — Type

Bridges CP.NonOverlappingOrthotopes to CP.Disjunction of linear inequations (MOI.LessThan{T}).

Variable number of constraints in the disjunction (two per dimension).

ConstraintProgrammingExtensions.Bridges.NonOverlappingOrthotopes2ConditionallyNonOverlappingOrthotopesBridge — Type

Bridges CP.NonOverlappingOrthotopes{CP.UNCONDITIONAL_NONVERLAPPING_ORTHOTOPES} to CP.NonOverlappingOrthotopes{CP.CONDITIONAL_NONVERLAPPING_ORTHOTOPES}.

ConstraintProgrammingExtensions.Bridges.Sort2MILPBridge — Type

Bridges CP.Sort to MILP constraints by adding O(n²) binary variables, with a transportation-like model.

Detailed model

Let x be the array to sort and y its sorted copy, both vectors having length n. This bridge handles the constraint [y..., x...]-in-CP.Sort(n).

Two sets of variables:

f[i, j]: real variable, i from 1 to n; the "flow" from the array to sort to the sorted copy, equal to the value of x[i] and y[j]

a[i, j]: binary variable, i from 1 to n; a[i, j] indicates whether the flow f[i, j] is nonzero

Constraints:

Flow coming from the array to sort x: $x_i = \sum_{j=1}^n f_{i,j} \qquad \forall i \in \{1, 2\dots n\}$

Flow going to the sorted array: $y_j = \sum_{i=1}^n f_{i,j} \qquad \forall j \in \{1, 2\dots n\}$

The flow from one value of the array to sort can only go to one element of the sorted array: $\sum_{i=1}^n a_{i,j} = 1 \qquad \forall j \in \{1, 2\dots n\}$

The flow to one value of the sorted array can only come from one element of the array to sort: $\sum_{j=1}^n a_{i,j} = 1 \qquad \forall i \in \{1, 2\dots n\}$

The flow f[i, j] is related to the binary variable a[i, j]: $U a_{i,j} \leq f_{i,j} \leq L a_{i,j} \qquad \forall (i,j) \in \{1, 2\dots n\}^2$

The array y must be sorted: $y_{j-1} \leq y_{j} \qquad \forall j \in \{2, 3\dots n\}$

ConstraintProgrammingExtensions.Bridges.Sort2SortPermutationBridge — Type

Bridges CP.Sort to CP.SortPermutation by adding index variables.

ConstraintProgrammingExtensions.Bridges.SortPermutation2AllDifferentBridge — Type

Bridges CP.SortPermutation to MOI.AllDifferent and CP.ElementVariableArray.

ConstraintProgrammingExtensions.Bridges.SlidingSum2LPBridge — Type

Bridges CP.SlidingSum to linear constraints.

ConstraintProgrammingExtensions.Bridges.SymmetricAllDifferent2AllDifferentInverseBridge — Type

Bridges CP.SymmetricAllDifferent to MOI.AllDifferent and CP.Inverse.

ConstraintProgrammingExtensions.Bridges.ValuePrecedence2ReificationBridge — Type

Bridges CP.ValuePrecedence to reification.

ConstraintProgrammingExtensions.Bridges.VectorDomain2MILPBridge — Type

Bridges CP.VectorDomain to MILP by adding one binary variable per possible combination.